refactor: modularize plugin system
