+++ src/lib/plugin-system.ts
@@ -3,52 +3,24 @@ import path from 'path';
 import { fileURLToPath } from 'url';
 import { Worker } from 'worker_threads';
 
-export interface Plugin {
-  id: string;
-  name: string;
-  onLoad?: (options?: Record<string, unknown>) => void | Promise<void>;
-  onEnable?: (options?: Record<string, unknown>) => void | Promise<void>;
-  onDisable?: () => void | Promise<void>;
-  enabled?: boolean;
-  options?: Record<string, unknown>;
-  permissions?: string[];
-  logs?: string[];
-  errors?: string[];
-  warning?: string;
-  error?: string;
-}
+import workerScript from './plugin/worker';
+import {
+  registry,
+  registerPlugin,
+  getPlugins,
+  enablePlugin,
+  disablePlugin,
+  updatePluginOptions,
+} from './plugin/registry';
+import type { Plugin } from './plugin/registry';
+import { loadState, setStateFilePath } from './plugin/state';
 
-interface PluginState {
-  enabled: boolean;
-  options?: Record<string, unknown>;
-}
-
-const registry = new Map<string, Plugin>();
 const workers = new Map<string, Worker>();
 const PLUGIN_TIMEOUT = 1000;
-let state: Record<string, PluginState> = {};
-let stateFilePath = path.join(process.cwd(), 'plugins', 'plugin-state.json');
-
-function loadState(): void {
-  if (fs.existsSync(stateFilePath)) {
-    try {
-      state = JSON.parse(fs.readFileSync(stateFilePath, 'utf8'));
-    } catch {
-      state = {};
-    }
-  } else {
-    state = {};
-  }
-}
-
-function saveState(): void {
-  fs.mkdirSync(path.dirname(stateFilePath), { recursive: true });
-  fs.writeFileSync(stateFilePath, JSON.stringify(state, null, 2));
-}
 
 function callWorker(worker: Worker, method: string, options?: Record<string, unknown>): Promise<void> {
   return new Promise((resolve, reject) => {
-    const listener = (msg: any) => {
+    const listener = (msg: { type: string; method: string; error?: string }) => {
       if (msg.type === 'result' && msg.method === method) {
         worker.off('message', listener);
         resolve();
@@ -62,53 +34,13 @@ function callWorker(worker: Worker, method: string, options?: Record<string, unk
   });
 }
 
-async function createWorker(filePath: string): Promise<{ worker: Worker; meta: { id: string; name: string; permissions?: string[] }; error?: string }>
-{
-  const worker = new Worker(
-    `import { parentPort, workerData } from 'worker_threads';
-import fs from 'fs';
-import { NodeVM } from 'vm2';
-import { transpileModule } from 'typescript';
-const file = workerData.file;
-let code = fs.readFileSync(file, 'utf8');
-if (file.endsWith('.ts')) {
-  code = transpileModule(code, { compilerOptions: { module: 1 } }).outputText;
-}
-const vm = new NodeVM({
-  console: 'redirect',
-  sandbox: {},
-  require: { external: false, builtin: [] },
-});
-vm.freeze(undefined, 'process');
-vm.on('console.log', (...args) => parentPort.postMessage({ type: 'log', data: args.join(' ') }));
-vm.on('console.error', (...args) => parentPort.postMessage({ type: 'consoleError', data: args.join(' ') }));
-let plugin;
-try {
-  plugin = vm.run(code, file).default;
-} catch (err) {
-  parentPort.postMessage({ type: 'loaded', error: String(err) });
-}
-if (plugin) {
-  parentPort.postMessage({ type: 'loaded', plugin: { id: plugin.id, name: plugin.name, permissions: plugin.permissions } });
-  parentPort.on('message', async (msg) => {
-    const { method, options, timeout } = msg;
-    try {
-      await Promise.race([
-        plugin[method]?.(options),
-        new Promise((_, reject) => setTimeout(() => reject(new Error('timeout')), timeout)),
-      ]);
-      parentPort.postMessage({ type: 'result', method });
-    } catch (err) {
-      parentPort.postMessage({ type: 'error', method, error: String(err) });
-    }
-  });
-}
-`,
-    { eval: true, type: 'module', workerData: { file: filePath } },
-  );
+async function createWorker(filePath: string): Promise<{ worker: Worker; meta: { id: string; name: string; permissions?: string[] }; error?: string }> {
+  const worker = new Worker(workerScript, { eval: true, type: 'module', workerData: { file: filePath } });
 
   return new Promise((resolve, reject) => {
-    const onMessage = (msg: any) => {
+    const onMessage = (
+      msg: { type: 'loaded'; plugin?: { id: string; name: string; permissions?: string[] }; error?: string },
+    ) => {
       if (msg.type === 'loaded') {
         worker.off('message', onMessage);
         resolve({ worker, meta: msg.plugin ?? { id: path.basename(filePath), name: path.basename(filePath) }, error: msg.error });
@@ -119,88 +51,11 @@ if (plugin) {
   });
 }
 
-export async function registerPlugin(plugin: Plugin): Promise<void> {
-  if (registry.has(plugin.id)) {
-    throw new Error(`Plugin with id "${plugin.id}" already registered`);
-  }
-
-  const persisted = state[plugin.id];
-  const entry: Plugin = {
-    ...plugin,
-    enabled: persisted?.enabled ?? plugin.enabled ?? false,
-    options: { ...plugin.options, ...persisted?.options },
-    permissions: plugin.permissions ?? [],
-    logs: plugin.logs ?? [],
-    errors: plugin.errors ?? [],
-  };
-
-  registry.set(plugin.id, entry);
-  try {
-    await plugin.onLoad?.(entry.options);
-  } catch (err) {
-    entry.error = String(err);
-  }
-  if (entry.enabled) {
-    try {
-      await plugin.onEnable?.(entry.options);
-    } catch (err) {
-      entry.error = String(err);
-    }
-  }
-  if (entry.permissions.length) {
-    entry.warning = `Requests permissions: ${entry.permissions.join(', ')}`;
-  }
-}
-
-export function getPlugins(): Plugin[] {
-  return Array.from(registry.values());
-}
-
-export async function enablePlugin(id: string): Promise<void> {
-  const plugin = registry.get(id);
-  if (plugin && !plugin.enabled) {
-    plugin.enabled = true;
-    try {
-      await plugin.onEnable?.(plugin.options);
-    } catch (err) {
-      plugin.error = String(err);
-    }
-    state[id] = { ...state[id], enabled: true, options: plugin.options };
-    saveState();
-  }
-}
-
-export async function disablePlugin(id: string): Promise<void> {
-  const plugin = registry.get(id);
-  if (plugin && plugin.enabled) {
-    plugin.enabled = false;
-    try {
-      await plugin.onDisable?.();
-    } catch (err) {
-      plugin.error = String(err);
-    }
-    state[id] = { ...state[id], enabled: false, options: plugin.options };
-    saveState();
-  }
-}
-
-export async function updatePluginOptions(
-  id: string,
-  options: Record<string, unknown>,
-): Promise<void> {
-  const plugin = registry.get(id);
-  if (plugin) {
-    plugin.options = options;
-    state[id] = { ...state[id], enabled: plugin.enabled ?? false, options };
-    saveState();
-  }
-}
-
 export async function loadPlugins(dir = path.join(process.cwd(), 'plugins')): Promise<void> {
   registry.clear();
   workers.forEach((w) => w.terminate());
   workers.clear();
-  stateFilePath = path.join(dir, 'plugin-state.json');
+  setStateFilePath(path.join(dir, 'plugin-state.json'));
   loadState();
   if (!fs.existsSync(dir)) return;
   const files = fs
@@ -226,11 +81,11 @@ export async function loadPlugins(dir = path.join(process.cwd(), 'plugins')): Pr
         continue;
       }
 
-      workers.set(entry.id, worker);
-      worker.on('message', (msg: any) => {
-        if (msg.type === 'log') entry.logs!.push(String(msg.data));
-        if (msg.type === 'consoleError') entry.errors!.push(String(msg.data));
-      });
+    workers.set(entry.id, worker);
+    worker.on('message', (msg: { type: 'log' | 'consoleError'; data: unknown }) => {
+      if (msg.type === 'log') entry.logs!.push(String(msg.data));
+      if (msg.type === 'consoleError') entry.errors!.push(String(msg.data));
+    });
 
       entry.onLoad = (opts) => callWorker(worker, 'onLoad', opts);
       entry.onEnable = (opts) => callWorker(worker, 'onEnable', opts);
@@ -248,7 +103,9 @@ export async function loadPlugins(dir = path.join(process.cwd(), 'plugins')): Pr
   }
 }
 
-// CLI execution: list loaded plugins
+export { registerPlugin, getPlugins, enablePlugin, disablePlugin, updatePluginOptions };
+export type { Plugin };
+
 if (process.argv[1] === fileURLToPath(import.meta.url)) {
   loadPlugins().then(() => {
     const loaded = getPlugins()
@@ -257,3 +114,4 @@ if (process.argv[1] === fileURLToPath(import.meta.url)) {
     console.log(`Loaded plugins: ${loaded || 'none'}`);
   });
 }
+
